+++ datastore/lib/oobium_datastore.dart
@@ -1,6 +1,7 @@
 export 'package:oobium_datastore/src/adapters.dart';
 export 'package:oobium_datastore/src/datastore.dart';
 export 'package:oobium_datastore/src/datastore/models.dart' show DataModel, DataId, HasMany, DataIndex, DataModelEvent;
+export 'package:oobium_datastore/src/errors.dart';
 
 // TODO remove ObjectId
 export 'package:objectid/objectid.dart';