+++ datastore_gen/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0-dev.5
+
+* BREAKING: remove inline options from schema
+* handle collections in adapter.decode
+* handle invalid types in adapter.encode
+* add @override annotation to model deleted method
+* updates for oobium_datastore 1.0.0-dev.5
+
 ## 1.0.0-dev.4
 
 * updates for oobium_datastore 1.0.0-dev.4