+++ routing/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0-dev.5
+
+* fix route matching
+
+## 1.0.0-dev.4
+
+no release
+
 ## 1.0.0-dev.3
 
 * add the home route during set(route) and setNewRoutePath(route)