+++ routing_gen/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.0-dev.5
+
+* version bump
+
+## 1.0.0-dev.4
+
+no release
+
+## 1.0.0-dev.3
+
+no release
+
 ## 1.0.0-dev.2
 
 * make RouteParser and RouterDelegate final fields of the generated Routes class